Removed lots of unnecessary casts to float of the Math methods. Also removed some other minor implicit casts

+++ Source/Core/Controls/SectorSlopeControl.cs
@@ -92,7 +92,7 @@ namespace CodeImp.DoomBuilder.Controls
 
 		#region ================== Methods
 
-		public void SetValues(float anglexy, float anglez, float offset, bool first) 
+		public void SetValues(double anglexy, double anglez, double offset, bool first) 
 		{
 			if(first) 
 			{
